Split App into multiple modules
