+++ integration/Setup.hs
@@ -111,7 +111,7 @@ testHooks hooks =
             dest
             ( unlines
                 [ "module RunAllTests where",
-                  "import Testlib.App",
+                  "import Testlib.Types",
                   "import Prelude",
                   unlines (map ("import qualified " <>) modules),
                   "allTests :: [(String, String, String, String, App ())]",
