+++ integration/test/Testlib/Cannon.hs
@@ -55,7 +55,12 @@ import qualified Network.WebSockets as WS
 import System.Random (randomIO)
 import System.Timeout (timeout)
 import Testlib.App
+import Testlib.Assertions
 import Testlib.Env
+import Testlib.HTTP
+import Testlib.JSON
+import Testlib.Printing
+import Testlib.Types
 
 data WebSocket = WebSocket
   { wsChan :: TChan Value,
