+++ integration/test/Testlib/Run.hs
@@ -1,6 +1,7 @@
 module Testlib.Run (main, mainI) where
 
 import Control.Concurrent
+import Control.Exception as E
 import Control.Monad
 import Data.Foldable
 import Data.Functor
@@ -9,9 +10,11 @@ import Data.Time.Clock
 import RunAllTests
 import System.Directory
 import System.Environment
-import Testlib.App
+import Testlib.Assertions
 import Testlib.Env
 import Testlib.Options
+import Testlib.Printing
+import Testlib.Types
 import Text.Printf
 import UnliftIO.Async
 
@@ -26,6 +29,21 @@ instance Semigroup TestReport where
 
 instance Monoid TestReport where
   mempty = TestReport 0 mempty
+
+runTest :: GlobalEnv -> App () -> IO (Maybe String)
+runTest ge action = do
+  env <- mkEnv ge
+  (runAppWithEnv env action $> Nothing)
+    `E.catches` [ E.Handler
+                    ( \(e :: AssertionFailure) -> do
+                        Just <$> printFailureDetails e
+                    ),
+                  E.Handler
+                    ( \(e :: SomeException) -> do
+                        putStrLn "exception handler"
+                        pure (Just (colored yellow (displayException e)))
+                    )
+                ]
 
 pluralise :: Int -> String -> String
 pluralise 1 x = x
