+++ integration/test/Testlib/Types.hs
@@ -1,0 +1,150 @@
+module Testlib.Types where
+
+import Control.Exception as E
+import Control.Monad.Catch
+import Control.Monad.Reader
+import qualified Data.Aeson as Aeson
+import qualified Data.Aeson.Encode.Pretty as Aeson
+import Data.ByteString (ByteString)
+import qualified Data.ByteString as BS
+import qualified Data.ByteString.Char8 as C8
+import qualified Data.ByteString.Lazy as L
+import qualified Data.CaseInsensitive as CI
+import Data.Functor
+import Data.List
+import qualified Data.Map as Map
+import qualified Data.Text as T
+import qualified Data.Text.Encoding as T
+import GHC.Records
+import GHC.Stack
+import qualified Network.HTTP.Client as HTTP
+import qualified Network.HTTP.Types as HTTP
+import Network.URI
+import Testlib.Env
+import Testlib.Printing
+
+data Response = Response
+  { jsonBody :: Maybe Aeson.Value,
+    body :: ByteString,
+    status :: Int,
+    headers :: [HTTP.Header],
+    request :: HTTP.Request
+  }
+
+instance HasField "json" Response (App Aeson.Value) where
+  getField response = maybe (assertFailure "Response has no json body") pure response.jsonBody
+
+showRequest :: HTTP.Request -> String
+showRequest r =
+  T.unpack (T.decodeUtf8 (HTTP.method r))
+    <> " "
+    <> uriToString id (HTTP.getUri r) ""
+
+showHeaders :: [HTTP.Header] -> String
+showHeaders r =
+  intercalate "\n" $
+    r <&> \(name, value) ->
+      C8.unpack (CI.original name) <> ": " <> C8.unpack value
+
+getRequestBody :: HTTP.Request -> Maybe BS.ByteString
+getRequestBody req = case HTTP.requestBody req of
+  HTTP.RequestBodyLBS lbs -> pure (L.toStrict lbs)
+  HTTP.RequestBodyBS bs -> pure bs
+  _ -> Nothing
+
+prettyResponse :: Response -> String
+prettyResponse r =
+  unlines $
+    concat
+      [ pure $ colored yellow "request: \n" <> showRequest r.request,
+        pure $ colored yellow "request headers: \n" <> showHeaders (HTTP.requestHeaders r.request),
+        case getRequestBody r.request of
+          Nothing -> []
+          Just b ->
+            [ colored yellow "request body:",
+              T.unpack . T.decodeUtf8 $ case Aeson.decode (L.fromStrict b) of
+                Just v -> L.toStrict (Aeson.encodePretty (v :: Aeson.Value))
+                Nothing -> b
+            ],
+        pure $ colored blue "response status: " <> show r.status,
+        pure $ colored blue "response body:",
+        pure $
+          ( T.unpack . T.decodeUtf8 $
+              case r.jsonBody of
+                Just b -> L.toStrict (Aeson.encodePretty b)
+                Nothing -> r.body
+          )
+      ]
+
+data AssertionFailure = AssertionFailure
+  { callstack :: CallStack,
+    response :: Maybe Response,
+    msg :: String
+  }
+
+instance Show AssertionFailure where
+  show (AssertionFailure _ _ msg) = "AssertionFailure _ _ " <> show msg
+
+instance Exception AssertionFailure where
+  displayException (AssertionFailure _ _ msg) = msg
+
+newtype App a = App {unApp :: ReaderT Env IO a}
+  deriving
+    ( Functor,
+      Applicative,
+      Monad,
+      MonadIO,
+      MonadMask,
+      MonadCatch,
+      MonadThrow,
+      MonadReader Env
+    )
+
+runAppWithEnv :: Env -> App a -> IO a
+runAppWithEnv e m = runReaderT (unApp m) e
+
+getServiceMap :: String -> App ServiceMap
+getServiceMap fedDomain = do
+  env <- ask
+  assertJust ("Could not find service map for federation domain: " <> fedDomain) (Map.lookup fedDomain (env.serviceMap))
+
+data AppFailure = AppFailure String
+
+instance Show AppFailure where
+  show (AppFailure msg) = msg
+
+instance Exception AppFailure where
+  displayException (AppFailure msg) = msg
+
+instance MonadFail App where
+  fail msg = assertFailure ("Pattern matching failure: " <> msg)
+
+assertFailure :: HasCallStack => String -> App a
+assertFailure msg =
+  forceList msg $
+    liftIO $
+      E.throw (AssertionFailure callStack Nothing msg)
+  where
+    forceList [] y = y
+    forceList (x : xs) y = seq x (forceList xs y)
+
+assertJust :: HasCallStack => String -> Maybe a -> App a
+assertJust _ (Just x) = pure x
+assertJust msg Nothing = assertFailure msg
+
+addFailureContext :: String -> App a -> App a
+addFailureContext msg = modifyFailureMsg (\m -> m <> "\nThis failure happend in this context:\n" <> msg)
+
+modifyFailureMsg :: (String -> String) -> App a -> App a
+modifyFailureMsg modMessage = modifyFailure (\e -> e {msg = modMessage e.msg})
+
+modifyFailure :: (AssertionFailure -> AssertionFailure) -> App a -> App a
+modifyFailure modifyAssertion action = do
+  env <- ask
+  liftIO
+    ( E.catch
+        (runAppWithEnv env action)
+        ( \(e :: AssertionFailure) ->
+            E.throw (modifyAssertion e)
+        )
+    )
